+++ src/main/java/org/prgms/springbootbasic/domain/voucher/VoucherPolicy.java
@@ -1,9 +1,5 @@
 package org.prgms.springbootbasic.domain.voucher;
 
-import java.util.UUID;
-
 public interface VoucherPolicy {
-    UUID getVoucherId();
-    long discount(long beforeDiscount);
-    long getDiscountDegree();
+    long discount(long beforeDiscount, long discountDegree);
 }
